+++ vins_estimator/src/feature_manager.h
@@ -90,6 +90,7 @@ class FeatureManager
     void removeBack();
     void removeFront(int frame_count);
     void removeOutlier();
+    void scaleDepth(double scale);
     list<FeaturePerId> feature;
     int last_track_num;